fix: remove facet-browser folder at move helpers to search-result folder and cleanup after
