+++ src/apps/search-result/facet-search.graphql
