+++ src/apps/search-result/helper.ts
@@ -1,9 +1,14 @@
-import { head } from "lodash";
-import { Filter } from "../../core/filter.slice";
+import { head, mapValues } from "lodash";
 import { getMaterialTypes } from "../../core/utils/helpers/general";
 import { Manifestation } from "../../core/utils/types/entities";
 import { ManifestationMaterialType } from "../../core/utils/types/material-type";
-import { FacetFieldEnum } from "../../core/dbc-gateway/generated/graphql";
+import {
+  FacetFieldEnum,
+  FacetResult,
+  FacetValue
+} from "../../core/dbc-gateway/generated/graphql";
+import { FacetOrigin, Filter, FilterItemTerm } from "../../core/filter.slice";
+import invalidSwitchCase from "../../core/utils/helpers/invalid-switch-case";
 
 export const mapFacetToFilter = (facet: FacetFieldEnum) => {
   switch (facet) {
@@ -35,27 +40,151 @@ export const getFirstMaterialTypeFromFilters = (
     : undefined;
 };
 
-export const formatSearchDisplayQuery = ({
-  q,
-  creator,
-  subject,
-  dk5,
-  t
-}: {
-  q?: string;
-  creator?: string | null;
-  subject?: string | null;
-  dk5?: string | null;
-  t: (key: string) => string;
-}): string => {
-  return [
-    q && q !== "*" ? q : null,
-    creator ? `${t("byAuthorText")}: ${creator}` : null,
-    subject ? `${t("facetSubjectsText")}: ${subject}` : null,
-    dk5 ? `${t("facetDk5Text")}: ${dk5}` : null
-  ]
-    .filter(Boolean)
-    .join("; ");
+export const allFacetFields = [
+  FacetFieldEnum.Materialtypesgeneral,
+  FacetFieldEnum.Materialtypesspecific,
+  FacetFieldEnum.Creators,
+  FacetFieldEnum.Year,
+  FacetFieldEnum.Mainlanguages,
+  FacetFieldEnum.Subjects,
+  FacetFieldEnum.Genreandform,
+  FacetFieldEnum.Fictionalcharacters,
+  FacetFieldEnum.Dk5,
+  FacetFieldEnum.Age,
+  FacetFieldEnum.Libraryrecommendation,
+  FacetFieldEnum.Generalaudience,
+  FacetFieldEnum.Lix,
+  FacetFieldEnum.Let,
+  FacetFieldEnum.Gameplatform,
+  FacetFieldEnum.Accesstypes,
+  FacetFieldEnum.Canalwaysbeloaned,
+  FacetFieldEnum.Fictionnonfiction,
+  FacetFieldEnum.Childrenoradults
+];
+
+export const formatFacetTerms = (filters: {
+  [key: string]: { [key: string]: FilterItemTerm };
+}) => {
+  return Object.keys(filters).reduce(
+    (acc, key) => ({
+      ...acc,
+      [key]: Object.keys(filters[key])
+    }),
+    {}
+  );
+};
+
+export const createFilters = (
+  facets: {
+    [key: string]: { [key: string]: FilterItemTerm };
+  },
+  branchIdList: string[]
+) => {
+  return {
+    ...formatFacetTerms(facets),
+    ...(branchIdList ? { branchId: branchIdList } : {})
+  };
+};
+
+export const FacetBrowserModalId = "facet-browser-modal";
+
+export function getAllFilterPathsAsString(
+  filterObject: Filter,
+  origin: FacetOrigin
+) {
+  const mappedFilterValues = mapValues(filterObject, (filter) => {
+    return Object.keys(filter).filter((term) => filter[term].origin === origin);
+  });
+
+  const filterNames = Object.keys(mappedFilterValues);
+  let allFilterPathsAsString = "";
+
+  filterNames.forEach((filterName) => {
+    mappedFilterValues[filterName].forEach((filterValue) => {
+      if (allFilterPathsAsString !== "") {
+        allFilterPathsAsString = allFilterPathsAsString.concat(";");
+      }
+      allFilterPathsAsString = allFilterPathsAsString.concat(
+        `facet.${filterName}:${filterValue}`
+      );
+    });
+  });
+
+  return allFilterPathsAsString;
+}
+
+export const getFacetFieldTranslation = (name: FacetFieldEnum) => {
+  switch (name.toLowerCase()) {
+    case FacetFieldEnum.Accesstypes.toLowerCase():
+      return "facetAccessTypesText";
+    case FacetFieldEnum.Canalwaysbeloaned.toLowerCase():
+      return "facetCanAlwaysBeLoanedText";
+    case FacetFieldEnum.Childrenoradults.toLowerCase():
+      return "facetChildrenOrAdultsText";
+    case FacetFieldEnum.Creators.toLowerCase():
+      return "facetCreatorsText";
+    case FacetFieldEnum.Dk5.toLowerCase():
+      return "facetDk5Text";
+    case FacetFieldEnum.Fictionnonfiction.toLowerCase():
+      return "facetFictionNonfictionText";
+    case FacetFieldEnum.Fictionalcharacters.toLowerCase():
+      return "facetFictionalCharactersText";
+    case FacetFieldEnum.Genreandform.toLowerCase():
+      return "facetGenreAndFormText";
+    case FacetFieldEnum.Mainlanguages.toLowerCase():
+      return "facetMainLanguagesText";
+    case FacetFieldEnum.Materialtypesspecific.toLowerCase():
+      return "facetMaterialTypesSpecificText";
+    case FacetFieldEnum.Materialtypesgeneral.toLowerCase():
+      return "facetMaterialTypesGeneralText";
+    case FacetFieldEnum.Subjects.toLowerCase():
+      return "facetSubjectsText";
+    case FacetFieldEnum.Worktypes.toLowerCase():
+      return "facetWorkTypesText";
+    case FacetFieldEnum.Year.toLowerCase():
+      return "facetYearText";
+    case FacetFieldEnum.Gameplatform.toLowerCase():
+      return "facetGamePlatformText";
+    case FacetFieldEnum.Age.toLowerCase():
+      return "facetAgeText";
+    case FacetFieldEnum.Generalaudience.toLowerCase():
+      return "facetGeneralAudienceText";
+    case FacetFieldEnum.Lix.toLowerCase():
+      return "facetLixText";
+    case FacetFieldEnum.Let.toLowerCase():
+      return "facetLetText";
+    case FacetFieldEnum.Libraryrecommendation.toLowerCase():
+      return "facetLibraryRecommendationText";
+    default:
+      return invalidSwitchCase<string>(name as never);
+  }
+};
+
+type FacetMap = { [key: string]: FacetValue };
+// createFacetsMap generates a map for quick lookup and enhanced search
+// capabilities by combining facet name and term into a single string key.
+// Structure: { [facetName:termKey]: FacetValue }
+// Example Key: 'fictionalCharacters:Batman'
+export const createFacetsMap = (facets: FacetResult[]): FacetMap => {
+  return facets.reduce((acc: FacetMap, facet: FacetResult) => {
+    const newAcc = { ...acc };
+
+    facet.values.forEach((value) => {
+      const combinedKey = `${facet.name}:${value.key}`;
+      newAcc[combinedKey] = value;
+    });
+
+    return newAcc;
+  }, {});
+};
+
+export const findTermInFacetMap = (
+  facetName: string,
+  termName: string,
+  facetMap: FacetMap
+): FacetValue => {
+  const key = `${facetName}:${termName}`;
+  return facetMap[key];
 };
 
 export default {};
