+++ src/apps/search-result/useFilterHandler.tsx
@@ -15,10 +15,9 @@ import {
   setQueryParametersInUrl
 } from "../../core/utils/helpers/url";
 import { FacetFieldEnum } from "../../core/dbc-gateway/generated/graphql";
-import { mapFacetToFilter } from "./helper";
+import { getAllFilterPathsAsString, mapFacetToFilter } from "./helper";
 import { useEventStatistics } from "../../core/statistics/useStatistics";
 import { statistics } from "../../core/statistics/statistics";
-import { getAllFilterPathsAsString } from "../../components/facet-browser/helper";
 
 const useFilterHandler = () => {
   const { track } = useEventStatistics();
