+++ src/components/facet-line/FacetLine.tsx
@@ -2,10 +2,10 @@ import React, { memo } from "react";
 import { useIntelligentFacetsQuery } from "../../core/dbc-gateway/generated/graphql";
 import FacetLineSelected from "./FacetLineSelected";
 import FacetLineFilters from "./FacetLineFilters";
-import { createFilters } from "../facet-browser/helper";
 import useGetSearchBranches from "../../core/utils/branches";
 import FacetLineFiltersSkeleton from "./FacetLineFiltersSkeleton";
 import useFilterHandler from "../../apps/search-result/useFilterHandler";
+import { createFilters } from "../../apps/search-result/helper";
 
 type FacetLineProps = {
   q: string;
