+++ src/components/facet-line/FacetLineSelected.tsx
@@ -2,9 +2,9 @@ import React, { useEffect, useRef } from "react";
 import useFilterHandler from "../../apps/search-result/useFilterHandler";
 import ButtonTag from "../Buttons/ButtonTag";
 import { useText } from "../../core/utils/text";
-import { FacetBrowserModalId } from "../facet-browser/helper";
 import IconFilter from "@danskernesdigitalebibliotek/dpl-design-system/build/icons/basic/icon-filter.svg";
 import { useModalButtonHandler } from "../../core/utils/modal";
+import { FacetBrowserModalId } from "../../apps/search-result/helper";
 
 const FacetLineSelected = () => {
   const { filters, removeFromFilter } = useFilterHandler();
